v2/backend/pyproject.toml
[tool.poetry]
name = "banking-v2"
version = "0.1.0"
description = "Personal banking analytics API"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = ">=0.115.0"
uvicorn = {version = ">=0.30.0", extras = ["standard"]}
sqlalchemy = ">=2.0.0"
psycopg2-binary = ">=2.9.0"
alembic = ">=1.13.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
python-jose = {version = ">=3.3.0", extras = ["cryptography"]}
bcrypt = ">=4.0.0"
python-multipart = ">=0.0.9"
apscheduler = ">=3.10.0"
scikit-learn = ">=1.5.0"
joblib = ">=1.4.0"
pandas = ">=2.2.0"
"pdfminer.six" = ">=20231228"
beautifulsoup4 = ">=4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
httpx = ">=0.27.0"
black = ">=24.0.0"
mypy = ">=1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["jose.*", "pdfminer.*", "bs4.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
